+++ docs/storybook/package.json
@@ -45,12 +45,8 @@
     "@storybook/addon-essentials": "^8.6.14",
     "@storybook/addon-interactions": "^8.6.14",
     "@storybook/addon-links": "^9.0.9",
-    "@storybook/blocks": "^8.6.14",
-    "@storybook/manager-api": "^8.6.14",
     "@storybook/react": "^9.0.9",
     "@storybook/react-vite": "^9.0.9",
-    "@storybook/test": "^8.6.14",
-    "@storybook/theming": "^8.6.14",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@typescript-eslint/eslint-plugin": "~8.11.0",
