+++ docs/storybook/src/AppUiStory.tsx
@@ -10,7 +10,7 @@ import {
   Primary,
   Subtitle,
   Title,
-} from "@storybook/blocks";
+} from "@storybook/addon-docs/blocks";
 import {
   AppNotificationManager,
   ConfigurableUiContent,
