+++ docs/storybook/src/Introduction.mdx
@@ -1,4 +1,4 @@
-import { Meta } from "@storybook/blocks";
+import { Meta } from "@storybook/addon-docs/blocks";
 import Code from "../assets/code-brackets.svg";
 import Direction from "../assets/direction.svg";
 import Flow from "../assets/flow.svg";
