+++ docs/storybook/src/frontstage/Frontstage.stories.tsx
@@ -3,7 +3,7 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 import type { Meta, StoryObj } from "@storybook/react";
-import { fireEvent, within } from "@storybook/test";
+import { fireEvent, within } from "storybook/test";
 import {
   StagePanelLocation,
   StagePanelSection,
