+++ docs/storybook/src/hooks/useIsBackstageOpen.mdx
@@ -1,4 +1,4 @@
-import { Meta, Story } from "@storybook/blocks";
+import { Meta, Story } from "@storybook/addon-docs/blocks";
 
 import * as stories from "./useIsBackstageOpen.stories";
